+++ src/components/SignInSignUpModal/RegisterLoginHeader/RegisterLoginHeader.tsx
@@ -2,17 +2,11 @@ import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import Switch from '@mui/material/Switch';
 import Typography from '@mui/material/Typography';
-import { ChangeEvent } from 'react';
+import { RegisterLoginHeaderProps } from 'src/models/profile';
 
 import styles from '/src/components/SignInSignUpModal/RegisterLoginHeader/RegisterLoginHeader.module.css';
 
-interface Props {
-  checked: boolean;
-  onChange: (event: ChangeEvent<HTMLInputElement>) => void;
-  onClick: () => void;
-}
-
-function RegisterLoginHeader({ checked, onChange, onClick }: Props): JSX.Element {
+function RegisterLoginHeader({ checked, onChange, onClick }: RegisterLoginHeaderProps): JSX.Element {
   return (
     <Box className={styles.registerLoginHeader__Wrapper}>
       <Typography
