--- src/const.js
@@ -1,2 +1,0 @@
-export const REQUIRED_ERROR = 'This field is required.';
-export const INVALID_EMAIL_ERROR = 'Invalid email format.';
