--- src/schemas/LoginInterface.d.ts
@@ -1,4 +1,0 @@
-export interface LoginInterface {
-  email: string;
-  password: string;
-}
