--- src/schemas/ProfilePageInterface.d.ts
@@ -1,5 +1,0 @@
-import { RegisterInterface } from './RegisterInterface';
-
-export interface ProfilePageInterface extends RegisterInterface {
-  email?: string;
-}
