--- src/schemas/RegisterInterface.d.ts
@@ -1,5 +1,0 @@
-import { LoginInterface } from './LoginInterface';
-
-export interface RegisterInterface extends LoginInterface {
-  name: string;
-}
